Migrate multichapter from ffmpeg demuxer to mp4-merge
Fix #11
