+++ Cargo.toml
@@ -19,6 +19,7 @@ filetime = "0.2.20"
 fs_extra = "1.3.0"
 log = "0.4.17"
 merkle_hash = "3.6.1"
+mp4-merge = "0.1.7"
 normpath = "1.1.0"
 simplelog = "0.12.1"
 uuid = { version = "1.3.0", features = ["v4"] }
