+++ tests/data/expected_output/GoPro_7322.mp4
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b72216e88c86f5a143c13a3df0926b66d9039550d5edf7d259bacaac44225870
-size 652853591
+oid sha256:668bbc04356ba4d904f9b40114d221d1e059d5adcc5e12851fccae2ae0452eaa
+size 652857129
