+++ tests/data/expected_output/GoPro_7329.mp4
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:67401b0fa2219bebb8813f7718952cfe13eac579298a17c366905bc55b2e8bfe
-size 381247351
+oid sha256:3f52aa18e71d4a139f8e67cd0a6c051f136ff1f3ccb48f37f467b77bfc1ef3b9
+size 381249333
